requirements.txt
Pillow
requests
beautifulsoup4
python-slugify